Changes in MovableGameObjects and GameObjects - GameObject constructer now takes a filepath and has a makeSprite method
